requirements.txt
mcp
google-genai
prompt_toolkit
rich
mcp
python-dotenv